Hate.
Let me tell you how much I've come to hate PHP since I began to use it.

There are 2,566 commits in this repository.  If the word hate was hidden in
each revision of each file in each commit ten thousand times over it would not
equal one one-billionth of the hate I feel for PHP at this micro-instant.

Hate.

Hate.