Move most things to the Container model
Some thought still needs to be given to how we move Container around in
some situations - it's in global scope at the moment for controllers to
get at (eww), and quite a lot of classes are now linked to
ContainerSubject.
However, this does make it far easier to simply override the services
provided by ContainerSubject with mock objects in order to perform
tests.