Workaround for limit=X00 bug = more music variety.
Like Heart, but actually more music variety. Last.fm has a bug in their API still, which is resulting in weird results for the Worldwide Top Chart. Upping the limit a small amount will fix this, and give a little more variety in the playlists. Groups still not working, so if those are looking like they will never come back we can think about ditching those.