It is worse than I thought...
It is actually every multiple of 10. Changing again, this time to 149, which I have checked properly (I thought I had checked 150, but clearly I must have either misseen, or it just broke.) The upshot of these APIs (when they work) is that they are no longer calculated just weekly, so the charts we have should flow more as people listen to their music throughout the week, so suddenly popular songs are reflected instantly, not a full week after they become popular.