Last.fm is terrible!!
At some point I will probably write something to utilise Spotify's API instead of the mostly broken one of Last.fm. In the meantime, this will make the bits that do work work more like they should. Basically, it is a redo of the previous patch as things are still weird with multiples of 10 in the API and Last.fm seem like they will never fix anything (least of all if it is reported).