Don't try to tracklist demos
Demos have the show_season_id of '0', so in the (unusual) case where
there's a demo going on at the same time as a show (e.g. OB with demo in
S1) don't try to tracklist what they play, nor try to add the tracks
played to the demo instead.

Fixes #573