Update JS
Everything in here is from my understanding (which may be very wrong)

- Added extra glyphicons
Since the message.type is an index for which icon to use, I had to add
icons for index 4 and 5, as well as the desired retweet icon.

- Added extra locationNames
again added extra names to accompany the icons.

- Added if-statement for dialog box
I have added an if statement so tweets would be formatted with the
sender info.