Moving show subtypes into MyRadio, part 2 - logic (#818)
* Show Colours part 1 - database schema

* Add foreign key, also change default.

We're using the ID for "regular" as the default subtype. This DOES mean that non-show shows (ha!) will have a subtype, but that's okay because it shouldn't be displayed publicly anyway.

* Use a join table after all

Upon further reflection, it might actually make sense for show subtypes to change between seasons (or even during one). So move to a join table, on either show or show_season.

* Add the show subtype for all newly created shows.

Note that show edits aren't yet accounted for because of backfilling (or lack thereof).

* Refactoring

* Don't store colours in the database - also forgot about events

* I'm a moron.

* Cleaning up

* Sync with schema changes from Part 1

* Update "can't upload" message, also fix phpcs break

* Unbreak.

I'm an idiot soz

* Add my shitty migration script