[Experiment] Forcibly write-back objects to cache
The cache write-back mechanism currently relies on
ServiceAPI::__destruct() being called. I have a theory that some of the
podcast caching weirdness is due it not being called, possibly because
something somewhere leaks a reference to it in MyRadioDaemon.

Try forcibly writing it back to the cache to maintain integrity. It's a
hack, but then again relying on destructors being called was bound to go
well.